requirements.txt
fastapi
mysql-connector
